# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MU Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: size
          in: query
          schema: { type: string }
        - name: color
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, price_asc, price_desc, best_selling, most_reviewed] }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    put:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId, images, stock]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        stock: { type: integer }
        material: { type: ["string", "null"] }
        isNew: { type: boolean }
        isSale: { type: boolean }
        isFeatured: { type: boolean }
        rating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        soldCount: { type: integer }
        createdAt: { type: string }
    ProductList:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ProductInput:
      type: object
      required: [name, price, categoryId, stock]
      properties:
        name: { type: string }
        nameAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        price: { type: number }
        salePrice: { type: number }
        categoryId: { type: integer }
        images: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        stock: { type: integer }
        material: { type: string }
        isNew: { type: boolean }
        isSale: { type: boolean }
        isFeatured: { type: boolean }
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        userId: { type: integer }
        userName: { type: ["string", "null"] }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [productId, rating]
      properties:
        productId: { type: integer }
        rating: { type: integer }
        comment: { type: string }
...