# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MU Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  nameAr?: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  salePrice?: number;
  categoryId: number;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock: number;
  material?: string;
  isNew?: boolean;
  isSale?: boolean;
  isFeatured?: boolean;
}